refactor(indicator): move TitleInfo to indicatorMetadata, add getTitleInfo to metadata
- Move TitleInfo/TitleValueItem types from paneTitle.ts to indicatorMetadata.ts
- Add GetTitleInfoFn type and getTitleInfo optional field to IndicatorMetadata
- Add getTitleInfo to IndicatorDefinitionConfig for @Indicator decorator
- Update all 14 import sites to import TitleInfo from indicatorMetadata
- Keep backward-compat re-export in paneTitle.ts
